bug: remove concurrency
There is simply no good way to set up concurrency with Terraform without lots of ugly hacks, at least at the moment.

If it were possible to manage the underlying Cloud Run service, without Terraform attempting to create it, then this would be ideal. Unfortunately, that's not yet the case.